[24.12.2021] PageShop page
1. Refactor classes definition
2. Fix products fetch
