+++ src/components/AppRouter/AppRouter.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Routes ,Route, Navigate } from 'react-router-dom';
 import About from "../pageshop/About";
-import Pageshop from "../pageshop/Pageshop";
+import PageShop from "../pageshop/PageShop";
 import ProductItem from "../ProductItem";
 
 
@@ -10,7 +10,7 @@ const AppRouter = () => {
     return (
         <Routes>
             <Route path="/About" element={<About/>} />
-            <Route exact path="/Shop" element={<Pageshop/>}/>
+            <Route exact path="/Shop" element={<PageShop/>}/>
             <Route exact path="/Shop/:id" element={<ProductItem/>} />
             <Route path="*" element={<Navigate to ="/Shop" />}/>
         </Routes>
